8일차 과제 요구사항과 다른부분 수정

+++ src/javaDay8/NumberDTO.java
@@ -1,14 +1,14 @@
 package javaDay8;
 
 public class NumberDTO {
-    private double Num1;
-    private double Num2;
+    private int Num1;
+    private int Num2;
     private double result;
-    public double getNum1() {
+    public int getNum1() {
         return Num1;
     }
 
-    public double getNum2() {
+    public int getNum2() {
         return Num2;
     }
     public void setResult(double d){
